backend/requirements.txt
Django==3.2
python-dotenv==0.19.2
djangorestframework==3.11.2
PyJWT==2.1.0
djangorestframework-simplejwt==5.0.0
djoser==2.0.1
Pillow==9.2.0
drf-extra-fields==3.0.3
gunicorn==20.0.4
psycopg2-binary==2.8.6
django-utils-six==2.0
django-filter==21.1
python-decouple==3.5
django-extensions==3.1.5
